Performance improvements
On 143x1476x1476 volume the memory usage should be about 3-4GB, and it should
take less than a minute to execute (serially). Most of the time is IO.

This translates to >5x improvement to both serial runtime and memory usage.
Conservatively, I expect it's 1.5x faster and uses 3x less memory than the
production pipeline (assuming that its not currently using Neuroproof_minimal).

The output, according to h5diff, should not differ from the original code.

Main performance improvements:

1) quantize predictions into unsigned chars. There is a "hack" to account for
small differences in the quantization of the prediction and inverted prediction
channels.

2) RAG building (RagAlgs.h, and Stack.cpp). Sort-based algorithm. Basic idea is
write edges to an array, treat vertices as self-edges, and then sort the edge
list lexicographically. Then scan the edge list, and update the features
in-batch.

3) feature manager: do not compute features on individual pixels. Instead, just
histogram the quantized prediction values and compute the contribution of each
of the 255 possible values.

4) Classification: add early stopping criteria to vigra. This cuts
rag_agglomeration's runtime in half.